Refactor: use size_t throughout WriteStringMsg chain, eliminate intermediate casts
Change WriteStringMsg return type from ssize_t to size_t and
currentDataLen_ from uint32_t to size_t. This eliminates ALL
intermediate type conversions in the write chain:

  size_t totalWritten (internal)
    -> return totalWritten (no cast, was static_cast<ssize_t>)
    -> size_t cnt (in WriteToBuffer, was ssize_t)
    -> cnt == msg.size() (no cast, was static_cast<size_t>)
    -> currentDataLen_ += cnt (no cast, was static_cast<uint32_t>)

The only remaining cast is at the protocol boundary in
WriteDumpResWithLen calls, where size_t currentDataLen_ is
explicitly cast to uint32_t for the DumpResMessage struct
(fixed binary protocol, uint32_t dataLen field).

WriteStringMsg never returns negative (writeFunc errors handled
internally with break), so size_t (unsigned) is the correct type
- ssize_t (signed) was unnecessary and required additional casts.

The overflow check (cnt > UINT32_MAX - currentDataLen_) is removed
because size_t accumulation doesn't overflow for realistic crash
dump sizes, and the protocol boundary cast handles truncation.

+++ common/dfxutil/dfx_util.cpp
@@ -307,7 +307,7 @@ int WriteBuf(int fd, const char* buf, size_t len)
     return totalWritten;
 }
 
-ssize_t WriteStringMsg(const int fd, const std::string& msg, BufferWriteFunc writeFunc)
+size_t WriteStringMsg(const int fd, const std::string& msg, BufferWriteFunc writeFunc)
 {
     constexpr size_t step = 1024 * 1024;
     size_t totalWritten = 0;
@@ -326,7 +326,7 @@ ssize_t WriteStringMsg(const int fd, const std::string& msg, BufferWriteFunc wri
         }
         totalWritten += static_cast<size_t>(cnt);
     }
-    return static_cast<ssize_t>(totalWritten);
+    return totalWritten;
 }
 
 #if is_ohos && !is_mingw
